Revert "Update to 4.0.0-rc2"
This reverts commit 8a1a2e21565015f419394a6a1edbcaa878088aa2.

+++ Bot_NetCore/Listeners/GuildMessageListener.cs
@@ -56,15 +56,11 @@ namespace Bot_NetCore.Listeners
 
                                         var wClient = new WebClient();
                                         wClient.DownloadFile(attachment.Url, file);
-
-                                        var message = new DiscordMessageBuilder()
-                                            .WithContent("**Удаление сообщения**\n" +
-                                                        $"**Автор:** {e.Message.Author.Username}#{e.Message.Author.Discriminator} ({e.Message.Author.Id})\n" +
-                                                        $"**Канал:** {e.Channel}\n" +
-                                                        $"**Содержимое: ```{e.Message.Content}```**")
-                                            .WithFile(file);
-
-                                        await e.Guild.GetChannel(Bot.BotSettings.FulllogChannel).SendMessageAsync(message);
+                                        await e.Guild.GetChannel(Bot.BotSettings.FulllogChannel)
+                                            .SendFileAsync(file, "**Удаление сообщения**\n" +
+                                                              $"**Автор:** {e.Message.Author.Username}#{e.Message.Author.Discriminator} ({e.Message.Author.Id})\n" +
+                                                              $"**Канал:** {e.Channel}\n" +
+                                                              $"**Содержимое: ```{e.Message.Content}```**");
                                         File.Delete(file);
                                         return;
                                     }
@@ -168,11 +164,10 @@ namespace Bot_NetCore.Listeners
 
                 if (e.Message.Attachments.Count > 0 && !e.Message.Author.IsBot)
                 {
-                    var message = new DiscordMessageBuilder()
-                        .WithContent($"**Автор:** {e.Message.Author}\n" +
-                                    $"**Канал:**  {e.Message.Channel}\n" +
-                                    $"**Сообщение:** {e.Message.Id}\n" +
-                                    $"**Вложение:**\n");
+                    var message = $"**Автор:** {e.Message.Author}\n" +
+                                  $"**Канал:**  {e.Message.Channel}\n" +
+                                  $"**Сообщение:** {e.Message.Id}\n" +
+                                  $"**Вложение:**\n";
 
                     using (var wClient = new WebClient())
                     {
@@ -180,14 +175,23 @@ namespace Bot_NetCore.Listeners
                         var file = $"generated/attachments/{attachment.FileName}";
                         wClient.DownloadFile(attachment.Url, file);
 
-                        message.WithFile(file);
+                        //TODO: REMOVE THIS -> Автобан за гифку
+                        if ((new FileInfo(file).Length) == 1058939)
+                        {
+                            var member = await e.Guild.GetMemberAsync(e.Author.Id);
+                            await member.BanAsync(1);
+                        }
+                        //TODO
+                        else
+                        {
 
-                        var logMessage = await e.Guild.GetChannel(Bot.BotSettings.AttachmentsLog).SendMessageAsync(message);
-                        File.Delete(file);
+                            var logMessage = await e.Guild.GetChannel(Bot.BotSettings.AttachmentsLog).SendFileAsync(file, message);
+                            File.Delete(file);
 
-                        using (var fs = new FileStream("generated/attachments_messages.csv", FileMode.Append))
-                        using (var sw = new StreamWriter(fs))
-                            await sw.WriteLineAsync($"{e.Message.Id},{logMessage.Id}");
+                            using (var fs = new FileStream("generated/attachments_messages.csv", FileMode.Append))
+                            using (var sw = new StreamWriter(fs))
+                                await sw.WriteLineAsync($"{e.Message.Id},{logMessage.Id}");
+                        }
                     }
                 }
